convert to float

+++ model/vgg_mse.py
@@ -69,7 +69,7 @@ def ce_loss(outputs, labels):
     return F.cross_entropy(outputs, labels, reduction='mean')
 
 def mse_loss(outputs, labels):
-    return F.mse_loss(outputs, labels, reduction='mean')
+    return F.mse_loss(outputs.float(), labels, reduction='mean')
 
 def accuracy(outputs, labels, split=None, images_name=None, fd=None):
     """
